J28-PublishPostsInGroupFinal

+++ Jobsway2goMvc/Controllers/GroupsController.cs
@@ -20,13 +20,11 @@ namespace Jobsway2goMvc.Controllers
             _context = context;
         }
 
-        // GET: Groups
         public async Task<IActionResult> Index()
         {
               return View(await _context.Groups.ToListAsync());
         }
 
-        // GET: Groups/Details/5
         public async Task<IActionResult> Details(int? id)
         {
             if (id == null || _context.Groups == null)
@@ -35,28 +33,30 @@ namespace Jobsway2goMvc.Controllers
             }
 
             var @group = await _context.Groups
+                .Include(g => g.Posts)
                 .FirstOrDefaultAsync(m => m.Id == id);
             if (@group == null)
             {
                 return NotFound();
             }
-
-            return View(@group);
+            ViewBag.GroupId = id;
+            return View(group);
+        }
+        public IActionResult DetailsPost(int id)
+        {
+            return RedirectToAction("Details", "Post", new { id = id });
         }
 
-        // GET: Groups/Create
         public IActionResult Create()
         {
             return View();
         }
 
-        // POST: Groups/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("Id,Name")] Group @group)
         {
+            ModelState.Remove("Posts");
             if (ModelState.IsValid)
             {
                 _context.Add(@group);
@@ -66,7 +66,6 @@ namespace Jobsway2goMvc.Controllers
             return View(@group);
         }
 
-        // GET: Groups/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
             if (id == null || _context.Groups == null)
@@ -82,9 +81,6 @@ namespace Jobsway2goMvc.Controllers
             return View(@group);
         }
 
-        // POST: Groups/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("Id,Name")] Group @group)
@@ -93,7 +89,7 @@ namespace Jobsway2goMvc.Controllers
             {
                 return NotFound();
             }
-
+            ModelState.Remove("Posts");
             if (ModelState.IsValid)
             {
                 try
@@ -117,7 +113,11 @@ namespace Jobsway2goMvc.Controllers
             return View(@group);
         }
 
-        // GET: Groups/Delete/5
+        public IActionResult EditPost(int id)
+        {
+            return RedirectToAction("Edit", "Post", new { id = id });
+        }
+
         public async Task<IActionResult> Delete(int? id)
         {
             if (id == null || _context.Groups == null)
@@ -135,23 +135,31 @@ namespace Jobsway2goMvc.Controllers
             return View(@group);
         }
 
-        // POST: Groups/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            if (_context.Groups == null)
+            if (_context.Groups == null || _context.Posts == null)
             {
-                return Problem("Entity set 'ApplicationDbContext.Groups'  is null.");
+                return Problem("Entity set 'ApplicationDbContext.Groups' or 'ApplicationDbContext.Posts'  is null.");
             }
             var @group = await _context.Groups.FindAsync(id);
             if (@group != null)
             {
+
+                var posts = _context.Posts.Where(p => p.GroupId == id);
+                _context.Posts.RemoveRange(posts);
+
                 _context.Groups.Remove(@group);
             }
-            
+
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
+        }
+
+        public IActionResult DeletePost(int id)
+        {
+            return RedirectToAction("Delete", "Post", new { id = id });
         }
 
         private bool GroupExists(int id)
